requirements.txt
black
discord.py
gevent
emoji
flake8
flask
flask-discord
Flask-Limiter
nest-asyncio
pynacl
python-dotenv
requests
youtube_dl
